+++ src/entry_point/entry_point.c
@@ -1,64 +1,59 @@
 #include "base.h"
 #include "display.h"
 #include "ctrl.h"
+#include "input.h"
 
+#include "input/input_kernel.h"
 #include "interrupt/interrupt.h"
 #include "acpi/acpi.h"
-
-static
-u32 print_string(const u32 x, const u32 y, const char *text) {
-    u32 cursor = x;
-    while(*text) {
-        os_display_set_character(cursor, y, *text, OS_DISPLAY_White);
-        ++cursor;
-        ++text;
-    }
-
-    return cursor;
-}
-
-static
-void print_number(const u32 x, const u32 y, const u32 number) {
-    const OS_Display_Color color = OS_DISPLAY_Cyan;
-
-    if(number == 0) {
-        os_display_set_character(x, y, '0', color);
-        return;
-    }
-
-    u32 power = 0;
-
-    {
-        u32 value = number;
-        while(value > 0) {
-            value /= 10;
-            ++power;
-        }
-    }
-
-    {
-        u32 value = number;
-        u32 index = 1;
-        while(value > 0) {
-            os_display_set_character(x + power - index, y, value % 10 + '0', color);
-            value /= 10;
-            ++index;
-        }
-    }
-}
 
 static
 void app(void) {
     s32 width, height;
     os_display_get_resolution(&width, &height);
+    os_display_clear(' ', OS_DISPLAY_White);
 
-    u32 seconds_passed = 0;
+    u32 cursor_x = 0;
     while(!os_ctrl_exit_requested()) {
-        os_display_clear(' ', OS_DISPLAY_White);
-        const u32 cursor = print_string(0, 0, "Seconds passed: ");
-        print_number(cursor, 0, seconds_passed);
-        os_ctrl_sleep(1000000000);
-        ++seconds_passed;
+        OS_Input_Event event;
+        while(os_input_pop_event(&event)) {
+            if(event.kind == OS_INPUT_EVENT_KIND_Keyboard && event.data.keyboard.down) {
+                switch(event.data.keyboard.key_code) {
+                    case OS_INPUT_KEY_Backspace:
+                        if(cursor_x > 0) --cursor_x;
+                        os_display_set_character(cursor_x, 0, ' ', OS_DISPLAY_White);
+                        break;
+
+                    case OS_INPUT_KEY_Arrow_Left:
+                        if(cursor_x > 0) --cursor_x;
+                        break;
+
+                    case OS_INPUT_KEY_Arrow_Right:
+                        ++cursor_x;
+                    break;
+
+                    case OS_INPUT_KEY_Space:
+                        if(event.data.keyboard.right_alt_down) {
+                            os_input_set_keyboard_layout(OS_INPUT_KEYBOARD_LAYOUT_Standard_US);
+                            break;
+                        }
+
+                        os_display_set_character(cursor_x, 0, ' ', OS_DISPLAY_White);
+                        ++cursor_x;
+                        break;
+
+                    default:
+                        if(event.data.keyboard.ascii == 0) break;
+                        u8 drawn_ascii = event.data.keyboard.ascii;
+                        if(event.data.keyboard.ascii >= 'A' && event.data.keyboard.ascii <= 'Z' && !event.data.keyboard.shift_down) {
+                            drawn_ascii = drawn_ascii - 'A' + 'a';
+                        }
+                        os_display_set_character(cursor_x, 0, drawn_ascii, OS_DISPLAY_White);
+                        ++cursor_x;
+                }
+            }
+        }
+        os_ctrl_sleep(10000000);
     }
 }
 
@@ -68,6 +63,7 @@ void app(void) {
  */
 int kernel_entry_point(void) {
     interrupt_initialize();
+    input_initialize();
     acpi_initialize();
     app();
     os_ctrl_shut_down();