+++ CHANGELOG.md
@@ -4,3 +4,5 @@
 
 - Added boot loader that calls into custom kernel implementation [!1](https://github.com/surrealtm/NarOS/pull/1)
 - Added a simple VGA video mode display API [!2](https://github.com/surrealtm/NarOS/pull/2)
+- Added a control API for time keeping and power management [!7](https://github.com/surrealtm/NarOS/pull/7)
+- Added an input API for keyboard events [!12](https://github.com/surrealtm/NarOS/pull/12)